% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddm.R
\name{dchoice}
\alias{dchoice}
\title{Marginal probability of one of the two alternative choices in a DDM}
\usage{
dchoice(b, a, v, choice = 1)
}
\arguments{
\item{b}{Relative starting point, a value between 0 and 1}

\item{a}{Inter-barrier distance, a positive real}

\item{v}{Drift rate, negative or positive real}

\item{choice}{Binary, 0 or 1.}
}
\value{
The probability of first crossing the upper or lower barrier
}
\description{
From Equation 3 of Tuerlinckx et al 2001. They use \code{z} in their equation,
which is \code{b*a} using parameters here. Eq 3 gives probability of crossing
upper barrier. Probability of crossing lower barrier uses the same formula
but with b = 1-b and v = -v.
}
\examples{
dchoice(b= 0.5, a = 2, v = 0.1)
}
\references{
Tuerlinckx, F., Maris, E., Ratcliff, R., & De Boeck, P. (2001). A comparison
of four methods for simulating the diffusion process. Behavior Research
Methods, Instruments, & Computers, 33(4), 443-456.
}
