% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddm.R
\name{ddm_stan}
\alias{ddm_stan}
\title{Bayesian inference of a drift diffusion model}
\usage{
ddm_stan(formula, data, tau = 0.1, alpha_ub = 5, sigma = 2, ...)
}
\arguments{
\item{formula}{A formula of the kind \code{y + z ~ 1}}

\item{data}{A data frame with response times and response choices}

\item{tau}{The non-decision time}

\item{alpha_ub}{The upper bound on the uniform prior on alpha}

\item{sigma}{The standard deviation of the normal prior on delta}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. chains, cores)}
}
\value{
A stan model with MCMC samples
}
\description{
Bayesian inference of a drift diffusion model
}
\examples{
set.seed(10101)
sims_df <- simulate_ddm(1000, b = 0.75, a = 2, v = 0.5) \%>\%
    dplyr::mutate(time = time + 0.1)
ddm_stan(time + choice ~ 1, data = sims_df, cores = 4, refresh = 0)
}
