% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{random_walk}
\alias{random_walk}
\title{Random walk approximation to a single trajectory in Drift-Diffusion Model}
\usage{
random_walk(b, a, v, t_eps = 1e-04)
}
\arguments{
\item{b}{Starting point in terms of proportion of distance from 0 to a.}

\item{a}{Upper barrier (lower is 0)}

\item{v}{Drift rate}

\item{t_eps}{Time step size}
}
\description{
Simulates a random walk with step sizes that are the square root of a
discrete time step t_eps. As t_eps -> 0, this random walk converges to a
drift-diffusion model.
}
