% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddm.R
\name{simulate_ddm}
\alias{simulate_ddm}
\title{Simulate a drift diffusion model}
\usage{
simulate_ddm(n, b, a, v, t_eps = 1e-04)
}
\arguments{
\item{n}{Number of independent samples}

\item{b}{Starting point}

\item{a}{Inter-barrier distance}

\item{v}{Drift-rate}

\item{t_eps}{Time-step size}
}
\value{
A data-frame with n rows and two columns, one the choice (which
barrier crossed first) and the other the time of the crossing.
}
\description{
Using the random walk approximation method described by Tuerlinckx et al
2001.
}
\examples{
simulate_ddm(10, 0.5, 2, 1)
}
\references{
Tuerlinckx, F., Maris, E., Ratcliff, R., & De Boeck, P. (2001). A comparison
of four methods for simulating the diffusion process. Behavior Research
Methods, Instruments, & Computers, 33(4), 443-456.
}
