% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddm.R
\name{time_choice_density}
\alias{time_choice_density}
\title{Bivariate density of response time and choice in a drift-diffusion model}
\usage{
time_choice_density(t, b, a, v, choice = 1)
}
\arguments{
\item{t}{time}

\item{b}{Relative starting point}

\item{a}{Upper boundary}

\item{v}{Drift rate}

\item{choice}{A binary choice represented as 0 and 1}
}
\value{
A probability density
}
\description{
The bivariate distribution over response time and choice, when choice is 0,
is given by Eq (1) in the Appendix of Wabersich & Vandekerckhove (2014).
}
\details{
The bivariate distribution over response time and choice, when choice is 1,
is given by Eq (1) in the Appendix of Wabersich & Vandekerckhove (2014) by
replacing beta by 1 - beta, and replace delta by -delta.

Here, we implement this density using the method described by Navarro & Fuss
(2009). This first approximates the density in a DDM with drift 0 and
inter-barrier distance (a) of 1 for time t/a^2, and then re-scales this.
}
\examples{
time_choice_density(t=1, b=0.5, a = 2, v = 1)
}
\references{
Wabersich, D., & Vandekerckhove, J. (2014). The RWiener Package: an R Package
Providing Distribution Functions for the Wiener Diffusion Model. R Journal,
6(1).

Navarro, D. J., & Fuss, I. G. (2009). Fast and accurate calculations for
first-passage times in Wiener diffusion models. Journal of mathematical
psychology, 53(4), 222-230.
}
